# lazily load FreeTypeAbstraction