--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,4 +1,11 @@
-using UnicodePlots, ReferenceTests, Test
+using UnicodePlots, Test
+
+# Incorrect usage of LazyModules could easily bring up world age issues. We need to test
+# this before loading all other test dependencies -- because otherwise potential world age
+# issues get automatically resolved.
+include("world_age_issues.jl")
+
+using ReferenceTests
 using ReferenceTests: BeforeAfterFull
 using Dates: Date, Day
 import Random: seed!