--- a/test/world_age_issues.jl
+++ b/test/world_age_issues.jl
@@ -1,0 +1,7 @@
+@testset "no world age issue" begin
+    @testset "png_image" begin
+        tmpname = joinpath(tempdir(), "foo.png")
+        savefig(lineplot([0, 1]), tmpname)
+        @test isfile(tmpname)
+    end
+end