The story here:

- for users that don't export png files, they get 0.8s latency reduced
- for users that need, nothing is changed. -- the FreeTypeAbstraction loading is delayed to the first `png_image` call

Note: **LazyModules could possibly introduce world-age issues if not used appropriately**. Luckily here FreeTypeAbstraction is a perfect black box -- both `renderstring!` and `findfont` outputs something we already know, i.e., String, Matrix.

Before:

```julia
julia> @time using UnicodePlots
  1.676243 seconds (6.64 M allocations: 485.373 MiB, 4.00% gc time, 11.02% compilation time: 72% of which was recompilation)

julia> p = @time lineplot([0, 1]);
  0.429121 seconds (601.23 k allocations: 40.566 MiB, 2.89% gc time, 99.79% compilation time)

julia> p = @time lineplot([0, 1]);
  0.000054 seconds (99 allocations: 11.016 KiB)

julia> @time savefig(p, "foo.png");
  2.858724 seconds (6.03 M allocations: 417.951 MiB, 4.60% gc time, 74.08% compilation time: 18% of which was recompilation)

julia> @time savefig(p, "foo.png");
  0.269588 seconds (37.30 k allocations: 7.499 MiB)
```

After:

```julia
julia> @time using UnicodePlots
  0.880895 seconds (4.09 M allocations: 306.096 MiB, 3.19% compilation time: 5% of which was recompilation)

julia> p = @time lineplot([0, 1]);
  0.406931 seconds (585.60 k allocations: 39.381 MiB, 99.81% compilation time)

julia> p = @time lineplot([0, 1]);
  0.049308 seconds (38.73 k allocations: 2.581 MiB, 99.79% compilation time: 48% of which was recompilation)

julia> @time savefig(p, "foo.png");
  4.058094 seconds (9.49 M allocations: 651.377 MiB, 4.35% gc time, 68.28% compilation time: 5% of which was recompilation)

julia> @time savefig(p, "foo.png");
  0.642027 seconds (660.92 k allocations: 50.671 MiB, 5.24% gc time, 54.63% compilation time: 99% of which was recompilation)
```

---

This doesn't improve TTFX performance, only "reduces" the package loading latency.
It's perhaps a good idea to use this trick in large umbrella packages such as Plots and Makie so let me ping a few people here @jkrumbiegel @SimonDanisch @BeastyBlacksmith @lorenzoh @ChrisRackauckas