@t-bltg 2022-05-22
On second calls:

> julia> p = @time lineplot([0, 1]);
  0.049308 seconds (38.73 k allocations: 2.581 MiB, 99.79% compilation time: 48% of which was recompilation)

or

> julia> @time savefig(p, "foo.png");
  0.642027 seconds (660.92 k allocations: 50.671 MiB, 5.24% gc time, 54.63% compilation time: 99% of which was recompilation)

Why does it recompile ?