@johnnychen94 2022-05-22
Oh, interesting. I actually don't know much about the internals so I could be wrong. But let me try to explain it: is that world-age changes (every time we load a bunch of new methods, e.g., loading a package, the world age increases) trigger invalidation and thus recompilation.

```julia
julia> @time using UnicodePlots
  0.901052 seconds (4.09 M allocations: 306.083 MiB, 3.52% compilation time: 5% of which was recompilation)

julia> p = @time lineplot([0, 1]);
  0.418570 seconds (585.60 k allocations: 39.381 MiB, 99.80% compilation time)

julia> p = @time lineplot([0, 1]);
  0.000063 seconds (99 allocations: 11.016 KiB)

julia> @time savefig(p, "foo.png");
  4.212479 seconds (9.49 M allocations: 651.379 MiB, 4.44% gc time, 68.22% compilation time: 5% of which was recompilation)

julia> @time savefig(p, "foo.png");
  0.657460 seconds (660.92 k allocations: 50.671 MiB, 5.23% gc time, 55.35% compilation time: 99% of which was recompilation)

julia> @time savefig(p, "foo.png");
  0.280837 seconds (37.63 k allocations: 7.526 MiB)
```

The first call triggers the package (FreeTypeAbstraction) loading, invalidates a few methods, then the second call recompiles it. Things get normal when it comes to the third call -- because no more packages are loaded in the second call.