@t-bltg 2022-05-22
> Things get normal when it comes to the third call

Good, I was afraid the timings were showing a regression.