@johnnychen94 2022-05-22
The overhead is about 60-80ns in my machine, so should be nothing to worry about for common non-trivial operations.

```julia
julia> using Colors, BenchmarkTools, LazyModules

julia> @lazy import ImageCore = "a09fc81d-aa75-5fe9-8630-4744c3626534"
LazyModule(ImageCore)

julia> @btime rand(RGB);
  3.219 ns (0 allocations: 0 bytes)

julia> @btime rand(ImageCore.RGB);
  75.468 ns (1 allocation: 32 bytes)
```