@johnnychen94 2022-05-22
This PR becomes a perfect example to cut off the package loading latency using the LazyModules trick so let's wait for a few days to raise some discussion (or, Q&A). I'd like to see how other people think about this.