@johnnychen94 2022-05-22
The even interesting story here is that UnicodePlots itself is also a black box, so it's doable for downstream packages to lazily load UP as well, for instance, https://github.com/JuliaImages/DitherPunk.jl/pull/77 switches from Requires to LazyModules