@t-bltg 2022-05-26
@johnnychen94, it looks like recent addition of `LazyModules` fails `Plots.jl` with the `UnicodePlots` backend (`FileIO` related, export as png). Those tests are only ran on `ubuntu` which explain why `MacOS` and `Windows` are not failing.

See https://github.com/JuliaPlots/Plots.jl/runs/6595180678?check_suite_focus=true.

Can you take a look ?