@johnnychen94 2022-05-27
IIUC this is due to the same reason why LazyModules doesn't apply to MarchingCubes -- DataFormat is an alien type to the world-age until FileIO is loaded.

My best guess is that FileIO needs to provide a better black box interface, e.g., `save(:png, io, img)` using Symbol. Otherwise, we should probably just roll back to the eager version.